requirements.txt
python_version >= "3.10"
# TTS
gTTS
pydub
google_auth_oauthlib
google-api-python-client
openai
jsonschema
youtube-dl @ git+https://github.com/ytdl-org/youtube-dl.git
git+https://github.com/suno-ai/bark.git
sqlitedict
git+https://github.com/MTG/freesound-python.git
isodate
pathvalidate
fuzzywuzzy
python-Levenshtein
python-dotenv
feedparser
retry
supabase